+++ packages/api/src/services/db/checkDbConnectivity.ts
