+++ packages/api/src/services/db/checkDbConnectivity.unit.test.ts
@@ -1,8 +1,8 @@
 import { isError } from "@moj-bichard7/common/types/Result"
 
-import checkConnectivity from "./checkConnectivity"
+import checkDbConnectivity from "./checkDbConnectivity"
 
-describe("checkConnectivity", () => {
+describe("checkDbConnectivity", () => {
   const mockDatabase = {
     connection: jest.fn()
   }
@@ -14,7 +14,7 @@ describe("checkConnectivity", () => {
   it("should return true when the database returns a value", async () => {
     mockDatabase.connection.mockResolvedValueOnce([{ connected: 1 }])
 
-    const result = await checkConnectivity(mockDatabase as any)
+    const result = await checkDbConnectivity(mockDatabase as any)
 
     expect(isError(result)).toBe(false)
     expect(result).toBe(true)
@@ -23,7 +23,7 @@ describe("checkConnectivity", () => {
   it("should return false when the database throws an error", async () => {
     mockDatabase.connection.mockRejectedValueOnce(new Error("Connection Refused"))
 
-    const result = await checkConnectivity(mockDatabase as any)
+    const result = await checkDbConnectivity(mockDatabase as any)
 
     expect(result).toBe(false)
   })
@@ -32,7 +32,7 @@ describe("checkConnectivity", () => {
     jest.useFakeTimers()
     mockDatabase.connection.mockResolvedValueOnce(new Promise(() => {}))
 
-    const resultPromise = checkConnectivity(mockDatabase as any)
+    const resultPromise = checkDbConnectivity(mockDatabase as any)
     jest.advanceTimersByTime(2100)
 
     const result = await resultPromise
